refactor out old lpc1549
