+++ examples/test_usb_source_measure.py
@@ -400,7 +400,7 @@ class UsbSourceMeasureTest(BoardTop):
   def refinements(self) -> Refinements:
     return super().refinements() + Refinements(
       instance_refinements=[
-        (['mcu'], Lpc1549_48New),
+        (['mcu'], Lpc1549_48),
         (['reg_5v'], Tps54202h),
         (['reg_3v3'], Xc6209),
         (['reg_analog'], Ap2210),
